requirements.txt
aiodns==3.1.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
ccxt==4.1.72
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
distlib==0.3.7
filelock==3.13.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
pluggy==1.3.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyarrow==14.0.1
pycares==4.4.0
pycoingecko==3.1.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
pyproject-api==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
schedule==1.2.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
streamlit==1.28.2
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tox==4.11.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.6
validators==0.22.0
virtualenv==20.24.7
yarl==1.9.2
zipp==3.17.0
